Add files via upload
 You are in charge of n chain stores spread across the world.
Because of economic problems in your company, you need to cut costs. You have been asked to work out which
stores can share some resources such as couriers. You decide to find the two best candidates for merging by
choosing the closest pair of stores.
Assume that your program is given n and the location of each store in terms of latitude and longitude. Report
the closest pair of stores (in alphabetical order) and the distance between them to the nearest 0.1km.